+++ crates/zng-ext-window/src/vars.rs
@@ -21,16 +21,16 @@ use zng_view_api::{
 #[cfg(feature = "image")]
 use crate::FrameCaptureMode;
 #[cfg(feature = "image")]
-use zng_ext_image::Img;
+use zng_ext_image::ImageEntry;
 
 pub(super) struct WindowVarsData {
     chrome: Var<bool>,
     icon: Var<WindowIcon>,
     #[cfg(feature = "image")]
-    pub(super) actual_icon: Var<Option<Img>>,
+    pub(super) actual_icon: Var<Option<ImageEntry>>,
     cursor: Var<CursorSource>,
     #[cfg(feature = "image")]
-    pub(super) actual_cursor_img: Var<Option<(Img, PxPoint)>>,
+    pub(super) actual_cursor_img: Var<Option<(ImageEntry, PxPoint)>>,
     title: Var<Txt>,
 
     state: Var<WindowState>,
@@ -224,7 +224,7 @@ impl WindowVars {
     /// [`icon`]: Self::icon
     /// [`Img`]: zng_ext_image::Img
     #[cfg(feature = "image")]
-    pub fn actual_icon(&self) -> Var<Option<Img>> {
+    pub fn actual_icon(&self) -> Var<Option<ImageEntry>> {
         self.0.actual_icon.read_only()
     }
 
@@ -249,7 +249,7 @@ impl WindowVars {
     /// [`Img`]: zng_ext_image::Img
     /// [`PxPoint`]: zng_layout::unit::PxPoint
     #[cfg(feature = "image")]
-    pub fn actual_cursor_img(&self) -> Var<Option<(Img, PxPoint)>> {
+    pub fn actual_cursor_img(&self) -> Var<Option<(ImageEntry, PxPoint)>> {
         self.0.actual_cursor_img.read_only()
     }