+++ crates/zng/src/image.rs
@@ -74,7 +74,7 @@
 pub use zng_ext_image::{
     ColorType, IMAGE_RENDER, IMAGES, ImageCacheMode, ImageDataFormat, ImageDownscaleMode, ImageEntriesMode, ImageEntryKind, ImageFormat,
     ImageFormatCapability, ImageHash, ImageHasher, ImageLimits, ImageOptions, ImageRenderArgs, ImageSource, ImageSourceFilter, ImageVar,
-    Img, PathFilter, render_retain,
+    ImageEntry, PathFilter, render_retain,
 };
 
 #[cfg(feature = "http")]